requirements.txt
asn1crypto==0.24.0
backcall==0.1.0
beautifulsoup4==4.6.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
colorama==0.3.9
cryptography==1.9
decorator==4.3.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.0.5
django-braces==1.13.0
django-crispy-forms==1.7.2
django-debug-toolbar==1.9.1
django-filter==1.1.0
django-registration-redux==2.4
django-webpack-loader==0.6.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
gunicorn==19.9.0
html5lib==1.0.1
idna==2.6
ipdb==0.11
ipython==6.4.0
ipython-genutils==0.2.0
jdcal==1.4
jedi==0.12.0
jwt==0.5.3
lxml==4.2.1
Markdown==2.6.11
mysqlclient==1.3.12
numpy==1.14.3
openpyxl==2.5.3
pandas==0.22.0
parso==0.2.1
pickleshare==0.7.4
Pillow==5.2.0
prompt-toolkit==1.0.15
psutil==5.4.6
psycopg2==2.7.5
py==1.5.2
pycparser==2.18
Pygments==2.2.0
PyJWT==1.6.4
pytest==3.2.5
python-dateutil==2.7.3
pytz==2018.4
requests==2.18.4
selenium==3.13.0
simplegeneric==0.8.1
six==1.11.0
sqlparse==0.2.4
static3==0.7.0
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
whitenoise==3.3.1
